v3.0.1-2
3.0.1

- deny bug fixe
- leader class added
- add leader`s permissions without luckperms
- add luckperms sync time to config
- add luckperms sync luckperms ranks to jobs
- leader accesses added
- job setleader command added
- job invite <user> added (without jobname args just for job leader and members)
- bug fix
- optimization on member manager

+++ src/main/java/org/parham/seasonjob/data/member/MemberManager.java
@@ -41,17 +41,24 @@ public class MemberManager {
         }
 
         if (member.getJob() == null) {
-            try {
-                File folder = new File(SeasonJob.getInstance().getDataFolder(), "data");
-                File memFile = new File(folder, id + ".yml");
-                if (memFile.exists()) {
-                    memFile.delete();
-                }
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-
             members.remove(id);
+        }
+    }
+
+    public static void updateMembersDataFiles() {
+        File folder = new File(SeasonJob.getInstance().getDataFolder(), "data");
+
+        if (!members.isEmpty()) {
+            for (File memFile : folder.listFiles()) {
+
+                try {
+                    if (memFile.exists() && !members.containsKey(UUID.fromString(memFile.getName().replace(".yml", "")))) {
+                        memFile.delete();
+                    }
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
         }
     }
